google_cloud_functions/schedule_update_info/src/requirements.txt
google-api-python-client
google-auth
google-cloud-datastore
pytz
requests
